Add check in case content is moved already but something later errors out.

+++ app/models/workflow/file_group_delete.rb
@@ -92,8 +92,10 @@ class Workflow::FileGroupDelete < Workflow::Base
   protected
 
   def move_physical_content
-    FileUtils.mkdir_p(Settings.medusa.cfs.fg_delete_holding)
-    FileUtils.move(file_group.cfs_directory.absolute_path, holding_directory_path)
+    if Dir.exist?(file_group.cfs_directory.absolute_path)
+      FileUtils.mkdir_p(Settings.medusa.cfs.fg_delete_holding)
+      FileUtils.move(file_group.cfs_directory.absolute_path, holding_directory_path)
+    end
   end
 
   def restore_physical_content
